web/requirements.txt
amqp==2.2.1
backports.weakref==1.0rc1
billiard==3.5.0.3
bleach==1.5.0
celery==4.0.2
cycler==0.10.0
decorator==4.1.1
Django==1.11.3
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-filter==1.0.2
django-logtailer==1.0.1
django-material==1.0.0
djangorestframework==3.6.3
html5lib==0.9999999
kombu==4.0.2
Markdown==2.6.8
matplotlib==2.0.2
networkx==1.11
numpy==1.13.1
olefile==0.44
opencv-contrib-python==3.2.0.7
opencv-python==3.2.0.7
Pillow==4.1.1
protobuf==3.3.0
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
redis==2.10.5
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0
sklearn==0.0
tensorflow==1.2.1
vine==1.1.4
Werkzeug==0.12.2
